refactor: restructure project into feature modules
