+++ src/features/participants/api/participantService.ts
@@ -1,5 +1,5 @@
-import { supabase } from '../supabase';
-import { Participant, ParticipantCreate } from '../../types/participant.types';
+import { supabase } from '../../shared/api/supabase';
+import { Participant, ParticipantCreate } from '../types';
 
 export const participantService = {
   getAll: async () => {
